Separate out test mining clients
